refactoring to allow easier implementation of new subcommands
